requirements.txt
pika==1.1.0
python-dotenv
dsnparse==0.1.15
mysql-connector-python

